Editorial rephrasing; define empty array/object
Did an editorial pass over the document, rewriting most of the prose slightly.

Caught a place or two that was still referring to tags for object keys.

Sole non-editorial change was adding the final ambiguous case - a completely empty node. These are required to have an `(array)` or `(object)` tag to be valid.