KQL: swap around order of comparisons to reduce local ambiguity
generally in {A,E}BNF, if one literal is a prefix of another,
the longer one is specified first, otherwise the shorter one
will always match, either causing a spurious error or a parsing
ambiguity, unless using a parser architecture which automatically
resolved local ambiguities by collecting alternatives.

this is something the rest of the KDL and KQL specs do well,
but it seems this case slipped through the cracks.